GPU: Get rid of obsolete makefiles folder and move the opencl build headers that are still needed to utils
