--- GPU/GPUTracking/Standalone/makefiles/.gitignore
@@ -1,3 +1,0 @@
-.svn
-as
-callvc.bat
